fix: rename styles files from `.sx.ts` to `.styles.ts`
