+++ src/modules/auth/components/login/components/email-input.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box';
 import { BasicInput } from '../../shared';
 import { type FormInputProps } from '#/modules/auth/types';
 import { isInputFieldEmpty } from '#/modules/auth/utils';
-import { inputStyles as styles } from '#/modules/auth/styles/input.sx';
+import { inputStyles as styles } from '#/modules/auth/styles/input.styles';
 
 export function EmailInput(props: FormInputProps) {
   const { focusedField } = props;
