+++ src/modules/auth/components/login/components/password-input.tsx
@@ -2,7 +2,7 @@ import { useRef } from 'react';
 import Box from '@mui/material/Box';
 import { type FormInputProps } from '#/modules/auth/types';
 import { isInputFieldEmpty } from '#/modules/auth/utils';
-import { inputStyles as styles } from '#/modules/auth/styles/input.sx';
+import { inputStyles as styles } from '#/modules/auth/styles/input.styles';
 import { BasicInput } from '../../shared';
 
 export function PasswordInput(props: FormInputProps) {
